Remove custom hook
